[alloc @ alla@mint-20061127010511-deae7e50ed033835] Added version checking for a MySQL>= 4.0 feature, in the delete history functionality.

+++ history/lib/history.inc.php
@@ -168,8 +168,11 @@ class history extends db_entity {
     $db->query($query);
     $row = $db->row();
     if ($row["total"] >= (3*$this->max_to_display)) {
-      $query  = sprintf("DELETE FROM history WHERE personID = %d ORDER BY the_time LIMIT %d",$current_user->get_id(),$this->max_to_display, (2*$this->max_to_display));
-      $db->query($query);
+      // Can only use DELETE FROM .. ORDER BY syntax from mysql >= 4.0
+      if (version_compare($db->get_db_version(),"4.0",">=")) {
+        $query  = sprintf("DELETE FROM history WHERE personID = %d ORDER BY the_time LIMIT %d",$current_user->get_id(),$this->max_to_display, (2*$this->max_to_display));
+        $db->query($query);
+      }
     }
 
     $ignored_files = $this->get_ignored_files();
